make full-screen option available to all code editors
